refactor(purchase-order): 카탈로그 영역을 PurchaseOrderCatalog smart 컴포넌트로 분리
HqPurchaseOrderCreateView 비대 SFC 책임 단위 분리 4차 — 1000줄 위반 해결.
좌측 카탈로그(검색·필터·정렬·그룹·SKU 다중선택·재고 시뮬레이션·facet 칩)
~290줄 template + 관련 helper/computed ~150줄을 모두 자식이 자체 보유.
부모 view 1311 → 837 (-474, -36%). 누적 1625 → 837 (-49%).

- PurchaseOrderCatalog: store/composable 직접 import 하는 smart 컴포넌트.
  v-model 7개(keyword/vendorFilter/sortBy/shortageOnly/selectedSkus/
  collapsed/rowQuantities) + emit 3개(addSku/addGroup/scrollCart) +
  defineExpose 2개(focusSearch/scrollToSku) 로 외부 인터페이스 좁힘.
- 부모는 cart 도메인(추가·삭제·수량·창고변경·공급처전환·제출) 만 책임.
  scrollToCatalogSku / handleKeydown 단축키는 catalogRef 위임으로 변경.
- 부모에서 useFacets / usePurchaseOrderStockSim composable 호출 제거,
  PlusIcon / SearchIcon import 도 카탈로그 전용이라 제거.
